+++ scripts/rollback/rollback-3.0.5.sh
@@ -1,0 +1,231 @@
+#!/bin/bash
+# timestamp: 1644466374656
+
+npm dist-tag add @midwayjs/egg-layer@3.0.4 latest
+npm dist-tag add @midwayjs/egg-layer@3.0.5 beta
+npm dist-tag add @midwayjs/serverless-app@3.0.4 latest
+npm dist-tag add @midwayjs/serverless-app@3.0.5 beta
+npm dist-tag add @midwayjs/axios@2.14.6 latest
+npm dist-tag add @midwayjs/axios@3.0.5 beta
+npm dist-tag add @midwayjs/bootstrap@2.14.6 latest
+npm dist-tag add @midwayjs/bootstrap@3.0.5 beta
+npm dist-tag add @midwayjs/cache@2.14.6 latest
+npm dist-tag add @midwayjs/cache@3.0.5 beta
+npm dist-tag add @midwayjs/consul@2.14.6 latest
+npm dist-tag add @midwayjs/consul@3.0.5 beta
+npm dist-tag add @midwayjs/core@2.14.6 latest
+npm dist-tag add @midwayjs/core@3.0.4 beta
+npm dist-tag add @midwayjs/cos@2.14.6 latest
+npm dist-tag add @midwayjs/cos@3.0.5 beta
+npm dist-tag add @midwayjs/cross-domain@3.0.4 latest
+npm dist-tag add @midwayjs/cross-domain@3.0.5 beta
+npm dist-tag add @midwayjs/decorator@2.14.0 latest
+npm dist-tag add @midwayjs/decorator@3.0.4 beta
+npm dist-tag add @midwayjs/express-session@3.0.4 latest
+npm dist-tag add @midwayjs/express-session@3.0.5 beta
+npm dist-tag add @midwayjs/faas@2.14.6 latest
+npm dist-tag add @midwayjs/faas@3.0.5 beta
+npm dist-tag add @midwayjs/grpc@2.14.6 latest
+npm dist-tag add @midwayjs/grpc@3.0.5 beta
+npm dist-tag add @midwayjs/i18n@3.0.4 latest
+npm dist-tag add @midwayjs/i18n@3.0.5 beta
+npm dist-tag add @midwayjs/info@3.0.4 latest
+npm dist-tag add @midwayjs/info@3.0.5 beta
+npm dist-tag add @midwayjs/jwt@2.14.6 latest
+npm dist-tag add @midwayjs/jwt@3.0.5 beta
+npm dist-tag add @midwayjs/mock@2.14.6 latest
+npm dist-tag add @midwayjs/mock@3.0.5 beta
+npm dist-tag add @midwayjs/mongoose@2.14.6 latest
+npm dist-tag add @midwayjs/mongoose@3.0.5 beta
+npm dist-tag add @midwayjs/orm@2.14.6 latest
+npm dist-tag add @midwayjs/orm@3.0.5 beta
+npm dist-tag add @midwayjs/oss@2.14.6 latest
+npm dist-tag add @midwayjs/oss@3.0.5 beta
+npm dist-tag add @midwayjs/passport@2.14.6 latest
+npm dist-tag add @midwayjs/passport@3.0.5 beta
+npm dist-tag add @midwayjs/process-agent@2.14.6 latest
+npm dist-tag add @midwayjs/process-agent@3.0.5 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@2.14.6 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.0.5 beta
+npm dist-tag add @midwayjs/prometheus@2.14.6 latest
+npm dist-tag add @midwayjs/prometheus@3.0.5 beta
+npm dist-tag add @midwayjs/rabbitmq@2.14.6 latest
+npm dist-tag add @midwayjs/rabbitmq@3.0.5 beta
+npm dist-tag add @midwayjs/redis@2.14.6 latest
+npm dist-tag add @midwayjs/redis@3.0.5 beta
+npm dist-tag add @midwayjs/security@3.0.4 latest
+npm dist-tag add @midwayjs/security@3.0.5 beta
+npm dist-tag add @midwayjs/sequelize@2.14.6 latest
+npm dist-tag add @midwayjs/sequelize@3.0.5 beta
+npm dist-tag add @midwayjs/session@3.0.4 latest
+npm dist-tag add @midwayjs/session@3.0.5 beta
+npm dist-tag add @midwayjs/socketio@2.14.6 latest
+npm dist-tag add @midwayjs/socketio@3.0.5 beta
+npm dist-tag add @midwayjs/static-file@3.0.4 latest
+npm dist-tag add @midwayjs/static-file@3.0.5 beta
+npm dist-tag add @midwayjs/swagger@3.0.4 latest
+npm dist-tag add @midwayjs/swagger@3.0.5 beta
+npm dist-tag add @midwayjs/tablestore@2.14.6 latest
+npm dist-tag add @midwayjs/tablestore@3.0.5 beta
+npm dist-tag add @midwayjs/task@2.14.6 latest
+npm dist-tag add @midwayjs/task@3.0.5 beta
+npm dist-tag add @midwayjs/typegoose@2.14.6 latest
+npm dist-tag add @midwayjs/typegoose@3.0.5 beta
+npm dist-tag add @midwayjs/upload@3.0.4 latest
+npm dist-tag add @midwayjs/upload@3.0.5 beta
+npm dist-tag add @midwayjs/validate@3.0.4 latest
+npm dist-tag add @midwayjs/validate@3.0.5 beta
+npm dist-tag add @midwayjs/version@2.14.6 latest
+npm dist-tag add @midwayjs/version@3.0.5 beta
+npm dist-tag add @midwayjs/view-ejs@2.14.6 latest
+npm dist-tag add @midwayjs/view-ejs@3.0.5 beta
+npm dist-tag add @midwayjs/view-nunjucks@2.14.6 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.0.5 beta
+npm dist-tag add @midwayjs/view@2.14.6 latest
+npm dist-tag add @midwayjs/view@3.0.5 beta
+npm dist-tag add @midwayjs/express@2.14.6 latest
+npm dist-tag add @midwayjs/express@3.0.5 beta
+npm dist-tag add @midwayjs/koa@2.14.6 latest
+npm dist-tag add @midwayjs/koa@3.0.5 beta
+npm dist-tag add @midwayjs/web@2.14.6 latest
+npm dist-tag add @midwayjs/web@3.0.5 beta
+npm dist-tag add @midwayjs/ws@2.14.6 latest
+npm dist-tag add @midwayjs/ws@3.0.5 beta
+tnpm dist-tag add @midwayjs/egg-layer@3.0.4 latest
+tnpm dist-tag add @midwayjs/egg-layer@3.0.5 beta
+tnpm dist-tag add @midwayjs/serverless-app@3.0.4 latest
+tnpm dist-tag add @midwayjs/serverless-app@3.0.5 beta
+tnpm dist-tag add @midwayjs/axios@2.14.6 latest
+tnpm dist-tag add @midwayjs/axios@3.0.5 beta
+tnpm dist-tag add @midwayjs/bootstrap@2.14.6 latest
+tnpm dist-tag add @midwayjs/bootstrap@3.0.5 beta
+tnpm dist-tag add @midwayjs/cache@2.14.6 latest
+tnpm dist-tag add @midwayjs/cache@3.0.5 beta
+tnpm dist-tag add @midwayjs/consul@2.14.6 latest
+tnpm dist-tag add @midwayjs/consul@3.0.5 beta
+tnpm dist-tag add @midwayjs/core@2.14.6 latest
+tnpm dist-tag add @midwayjs/core@3.0.4 beta
+tnpm dist-tag add @midwayjs/cos@2.14.6 latest
+tnpm dist-tag add @midwayjs/cos@3.0.5 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.0.4 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.0.5 beta
+tnpm dist-tag add @midwayjs/decorator@2.14.0 latest
+tnpm dist-tag add @midwayjs/decorator@3.0.4 beta
+tnpm dist-tag add @midwayjs/express-session@3.0.4 latest
+tnpm dist-tag add @midwayjs/express-session@3.0.5 beta
+tnpm dist-tag add @midwayjs/faas@2.14.6 latest
+tnpm dist-tag add @midwayjs/faas@3.0.5 beta
+tnpm dist-tag add @midwayjs/grpc@2.14.6 latest
+tnpm dist-tag add @midwayjs/grpc@3.0.5 beta
+tnpm dist-tag add @midwayjs/i18n@3.0.4 latest
+tnpm dist-tag add @midwayjs/i18n@3.0.5 beta
+tnpm dist-tag add @midwayjs/info@3.0.4 latest
+tnpm dist-tag add @midwayjs/info@3.0.5 beta
+tnpm dist-tag add @midwayjs/jwt@2.14.6 latest
+tnpm dist-tag add @midwayjs/jwt@3.0.5 beta
+tnpm dist-tag add @midwayjs/mock@2.14.6 latest
+tnpm dist-tag add @midwayjs/mock@3.0.5 beta
+tnpm dist-tag add @midwayjs/mongoose@2.14.6 latest
+tnpm dist-tag add @midwayjs/mongoose@3.0.5 beta
+tnpm dist-tag add @midwayjs/orm@2.14.6 latest
+tnpm dist-tag add @midwayjs/orm@3.0.5 beta
+tnpm dist-tag add @midwayjs/oss@2.14.6 latest
+tnpm dist-tag add @midwayjs/oss@3.0.5 beta
+tnpm dist-tag add @midwayjs/passport@2.14.6 latest
+tnpm dist-tag add @midwayjs/passport@3.0.5 beta
+tnpm dist-tag add @midwayjs/process-agent@2.14.6 latest
+tnpm dist-tag add @midwayjs/process-agent@3.0.5 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@2.14.6 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.0.5 beta
+tnpm dist-tag add @midwayjs/prometheus@2.14.6 latest
+tnpm dist-tag add @midwayjs/prometheus@3.0.5 beta
+tnpm dist-tag add @midwayjs/rabbitmq@2.14.6 latest
+tnpm dist-tag add @midwayjs/rabbitmq@3.0.5 beta
+tnpm dist-tag add @midwayjs/redis@2.14.6 latest
+tnpm dist-tag add @midwayjs/redis@3.0.5 beta
+tnpm dist-tag add @midwayjs/security@3.0.4 latest
+tnpm dist-tag add @midwayjs/security@3.0.5 beta
+tnpm dist-tag add @midwayjs/sequelize@2.14.6 latest
+tnpm dist-tag add @midwayjs/sequelize@3.0.5 beta
+tnpm dist-tag add @midwayjs/session@3.0.4 latest
+tnpm dist-tag add @midwayjs/session@3.0.5 beta
+tnpm dist-tag add @midwayjs/socketio@2.14.6 latest
+tnpm dist-tag add @midwayjs/socketio@3.0.5 beta
+tnpm dist-tag add @midwayjs/static-file@3.0.4 latest
+tnpm dist-tag add @midwayjs/static-file@3.0.5 beta
+tnpm dist-tag add @midwayjs/swagger@3.0.4 latest
+tnpm dist-tag add @midwayjs/swagger@3.0.5 beta
+tnpm dist-tag add @midwayjs/tablestore@2.14.6 latest
+tnpm dist-tag add @midwayjs/tablestore@3.0.5 beta
+tnpm dist-tag add @midwayjs/task@2.14.6 latest
+tnpm dist-tag add @midwayjs/task@3.0.5 beta
+tnpm dist-tag add @midwayjs/typegoose@2.14.6 latest
+tnpm dist-tag add @midwayjs/typegoose@3.0.5 beta
+tnpm dist-tag add @midwayjs/upload@3.0.4 latest
+tnpm dist-tag add @midwayjs/upload@3.0.5 beta
+tnpm dist-tag add @midwayjs/validate@3.0.4 latest
+tnpm dist-tag add @midwayjs/validate@3.0.5 beta
+tnpm dist-tag add @midwayjs/version@2.14.6 latest
+tnpm dist-tag add @midwayjs/version@3.0.5 beta
+tnpm dist-tag add @midwayjs/view-ejs@2.14.6 latest
+tnpm dist-tag add @midwayjs/view-ejs@3.0.5 beta
+tnpm dist-tag add @midwayjs/view-nunjucks@2.14.6 latest
+tnpm dist-tag add @midwayjs/view-nunjucks@3.0.5 beta
+tnpm dist-tag add @midwayjs/view@2.14.6 latest
+tnpm dist-tag add @midwayjs/view@3.0.5 beta
+tnpm dist-tag add @midwayjs/express@2.14.6 latest
+tnpm dist-tag add @midwayjs/express@3.0.5 beta
+tnpm dist-tag add @midwayjs/koa@2.14.6 latest
+tnpm dist-tag add @midwayjs/koa@3.0.5 beta
+tnpm dist-tag add @midwayjs/web@2.14.6 latest
+tnpm dist-tag add @midwayjs/web@3.0.5 beta
+tnpm dist-tag add @midwayjs/ws@2.14.6 latest
+tnpm dist-tag add @midwayjs/ws@3.0.5 beta
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 3.0.4 => 3.0.5
+#  - @midwayjs/serverless-app: 3.0.4 => 3.0.5
+#  - @midwayjs/axios: 2.14.6 => 3.0.5
+#  - @midwayjs/bootstrap: 2.14.6 => 3.0.5
+#  - @midwayjs/cache: 2.14.6 => 3.0.5
+#  - @midwayjs/consul: 2.14.6 => 3.0.5
+#  - @midwayjs/core: 2.14.6 => 3.0.5
+#  - @midwayjs/cos: 2.14.6 => 3.0.5
+#  - @midwayjs/cross-domain: 3.0.4 => 3.0.5
+#  - @midwayjs/decorator: 2.14.0 => 3.0.5
+#  - @midwayjs/express-session: 3.0.4 => 3.0.5
+#  - @midwayjs/faas: 2.14.6 => 3.0.5
+#  - @midwayjs/grpc: 2.14.6 => 3.0.5
+#  - @midwayjs/i18n: 3.0.4 => 3.0.5
+#  - @midwayjs/info: 3.0.4 => 3.0.5
+#  - @midwayjs/jwt: 2.14.6 => 3.0.5
+#  - @midwayjs/mock: 2.14.6 => 3.0.5
+#  - @midwayjs/mongoose: 2.14.6 => 3.0.5
+#  - @midwayjs/orm: 2.14.6 => 3.0.5
+#  - @midwayjs/oss: 2.14.6 => 3.0.5
+#  - @midwayjs/passport: 2.14.6 => 3.0.5
+#  - @midwayjs/process-agent: 2.14.6 => 3.0.5
+#  - @midwayjs/prometheus-socket-io: 2.14.6 => 3.0.5
+#  - @midwayjs/prometheus: 2.14.6 => 3.0.5
+#  - @midwayjs/rabbitmq: 2.14.6 => 3.0.5
+#  - @midwayjs/redis: 2.14.6 => 3.0.5
+#  - @midwayjs/security: 3.0.4 => 3.0.5
+#  - @midwayjs/sequelize: 2.14.6 => 3.0.5
+#  - @midwayjs/session: 3.0.4 => 3.0.5
+#  - @midwayjs/socketio: 2.14.6 => 3.0.5
+#  - @midwayjs/static-file: 3.0.4 => 3.0.5
+#  - @midwayjs/swagger: 3.0.4 => 3.0.5
+#  - @midwayjs/tablestore: 2.14.6 => 3.0.5
+#  - @midwayjs/task: 2.14.6 => 3.0.5
+#  - @midwayjs/typegoose: 2.14.6 => 3.0.5
+#  - @midwayjs/upload: 3.0.4 => 3.0.5
+#  - @midwayjs/validate: 3.0.4 => 3.0.5
+#  - @midwayjs/version: 2.14.6 => 3.0.5
+#  - @midwayjs/view-ejs: 2.14.6 => 3.0.5
+#  - @midwayjs/view-nunjucks: 2.14.6 => 3.0.5
+#  - @midwayjs/view: 2.14.6 => 3.0.5
+#  - @midwayjs/express: 2.14.6 => 3.0.5
+#  - @midwayjs/koa: 2.14.6 => 3.0.5
+#  - @midwayjs/web: 2.14.6 => 3.0.5
+#  - @midwayjs/ws: 2.14.6 => 3.0.5